Handling provision issues making it run twice.
Fixes: https://github.com/freeipa/freeipa-pr-ci/issues/2
